Geom2D

Description

Basic 2 dimensional geometry functions.

Documentation

module Data.VectorSpace

module Data.Cross

pattern MV_Point :: MVector s (a, a) -> MVector s (Point a)

pattern V_Point :: Vector (a, a) -> Vector (Point a)

class AffineTransform a b | a -> b where

Methods

transform :: Transform b -> a -> a

data Polygon a

Constructors

Polygon [Point a]

data Line a

Constructors

Line (Point a) (Point a)

data Transform a

A transformation (x, y) -> (ax + by + c, dx + ey + d)

Constructors

Transform

Fields

xformA :: !a

xformB :: !a

xformC :: !a

xformD :: !a

xformE :: !a

xformF :: !a

type DPoint = Point Double

data Point a

Constructors

Point

Fields

pointX :: !a

pointY :: !a

($*) :: AffineTransform a b => Transform b -> a -> a infixr 5

Operator for applying a transformation.

inverse :: (Eq a, Fractional a) => Transform a -> Maybe (Transform a)

Calculate the inverse of a transformation.

lineEquation :: Floating t => Line t -> (t, t, t)

Return the parameters (a, b, c) for the normalised equation of the line: a*x + b*y + c = 0.

lineDistance :: Floating a => Line a -> Point a -> a

Return the signed distance from a point to the line. If the distance is negative, the point lies to the right of the line

closestPoint :: Fractional a => Line a -> Point a -> Point a

Return the point on the line closest to the given point.

lineIntersect :: (Ord a, Floating a) => Line a -> Line a -> a -> Maybe (Point a)

Calculate the intersection of two lines. If the determinant is less than tolerance (parallel or coincident lines), return Nothing.

vectorMag :: Floating a => Point a -> a

The lenght of the vector.

vectorMagSquare :: Floating a => Point a -> a

The lenght of the vector.

vectorAngle :: RealFloat a => Point a -> a

The angle of the vector, in the range (-pi, pi].

dirVector :: Floating a => a -> Point a

The unitvector with the given angle.

normVector :: Floating a => Point a -> Point a

The unit vector with the same direction.

(^.^) :: Num a => Point a -> Point a -> a

Dot product of two vectors.

vectorCross :: Num a => Point a -> Point a -> a

Cross product of two vectors.

vectorDistance :: Floating a => Point a -> Point a -> a

Distance between two vectors.

interpolateVector :: Num a => Point a -> Point a -> a -> Point a

Interpolate between two vectors.

rotateScaleVec :: Num a => Point a -> Transform a

Create a transform that rotates by the angle of the given vector and multiplies with the magnitude of the vector.

flipVector :: Num a => Point a -> Point a

reflect the vector over the X-axis.

turnAround :: Num a => Point a -> Point a

rotateVec :: Floating a => Point a -> Transform a

Create a transform that rotates by the angle of the given vector with the x-axis

rotate :: Floating s => s -> Transform s

Create a transform that rotates by the given angle (radians).

rotate90L :: Floating s => Transform s

Rotate vector 90 degrees left.

rotate90R :: Floating s => Transform s

Rotate vector 90 degrees right.

translate :: Num a => Point a -> Transform a

Create a transform that translates by the given vector.

idTrans :: Num a => Transform a

The identity transformation.