Geom2D.CubicBezier.Outline

Description

Offsetting bezier curves and stroking curves.

Documentation

bezierOffset

Arguments

:: CubicBezier Double
The curve

-> Double
Offset distance.

-> Maybe Int
maximum subcurves

-> Double
Tolerance.

-> [CubicBezier Double]
The offset curve

Calculate an offset path from the bezier curve to within tolerance. If the distance is positive offset to the left, otherwise to the right. A smaller tolerance may require more bezier curves in the path to approximate the offset curve

bezierOffsetPoint :: CubicBezier Double -> Double -> Double -> (DPoint, DPoint)