Geom2D.CubicBezier.Stroke

Documentation

penCircle :: Floating a => Pen a

A circular pen with unit radius.

pathToPen :: Floating a => ClosedPath a -> Pen a

Create a pen from a path. For predictable results the path should be convex.

penStrokeOpen :: Int -> Double -> Bool -> Pen Double -> OpenPath Double -> [ClosedPath Double]

penStrokeClosed :: Int -> Double -> Bool -> Pen Double -> ClosedPath Double -> [ClosedPath Double]

data Pen a

bezierOffset

Arguments

:: CubicBezier Double
The curve

-> Double
Offset distance.

-> Maybe Int
maximum subcurves

-> Double
Tolerance.

-> Bool
Calculate the curve faster but with more subcurves

-> [CubicBezier Double]
The offset curve

Calculate an offset path from the bezier curve to within tolerance. If the distance is positive offset to the left, otherwise to the right. A smaller tolerance may require more bezier curves in the path to approximate the offset curve