Geom2D.CubicBezier.Curvature

Documentation

curvature :: CubicBezier Double -> Double -> Double

Curvature of the Bezier curve. A negative curvature means the curve curves to the right.

radiusOfCurvature :: CubicBezier Double -> Double -> Double

Radius of curvature of the Bezier curve. This is the reciprocal of the curvature.

curvatureExtrema :: CubicBezier Double -> Double -> [Double]

Find extrema of the curvature, but not inflection points.

findRadius

Arguments

:: CubicBezier Double
the curve

-> Double
distance

-> Double
tolerance

-> [Double]
result

Find points on the curve that have a certain radius of curvature. Values to the left and to the right of the curve are returned.